% Generated by build script
\name{gdacs_event}
\alias{gdacs_event}
\title{Get a single GDACS event by event ID and type}
\usage{
gdacs_event()
}
\arguments{
\item{event_id}{Event ID number (e.g. "1000001")}
\item{event_type}{Event type code: "EQ", "TC", "FL", "VO", "DR", "WF"}
}
\value{tibble with one row of event details}
\description{Get a single GDACS event by event ID and type}
